@@ -5,13 +5,11 @@ from logging import getLogger
 from operator import attrgetter
 from typing import (
     TYPE_CHECKING,
-    Any,
     Dict,
     Iterable,
-    Iterator,
     List,
-    Optional,
     NamedTuple,
+    Optional,
     Sequence,
     Tuple,
     Type,
@@ -106,7 +104,7 @@ class Segment(NamedTuple):
 
     @classmethod
     @lru_cache(1024 * 16)
-    def _split_cells(cls, segment: "Segment", cut: int) -> Tuple["Segment", "Segment"]:  # type: ignore
+    def _split_cells(cls, segment: "Segment", cut: int) -> Tuple["Segment", "Segment"]:
 
         text, style, control = segment
         _Segment = Segment
@@ -141,6 +139,8 @@ class Segment(NamedTuple):
                     _Segment(before[: pos - 1] + " ", style, control),
                     _Segment(" " + text[pos:], style, control),
                 )
+
+        raise AssertionError("Will never reach here")
 
     def split_cells(self, cut: int) -> Tuple["Segment", "Segment"]:
         """Split segment in to two segments at the specified column.
@@ -689,39 +689,35 @@ class SegmentLines:
                 yield from line
 
 
-if __name__ == "__main__":
+if __name__ == "__main__":  # pragma: no cover
+    from rich.console import Console
+    from rich.syntax import Syntax
+    from rich.text import Text
 
-    if __name__ == "__main__":  # pragma: no cover
-        from rich.console import Console
-        from rich.syntax import Syntax
-        from rich.text import Text
+    code = """from rich.console import Console
+console = Console()
+text = Text.from_markup("Hello, [bold magenta]World[/]!")
+console.print(text)"""
 
-        code = """from rich.console import Console
-    console = Console()
     text = Text.from_markup("Hello, [bold magenta]World[/]!")
-    console.print(text)"""
 
-        text = Text.from_markup("Hello, [bold magenta]World[/]!")
+    console = Console()
 
-        console = Console()
-
-        console.rule("rich.Segment")
-        console.print(
-            "A Segment is the last step in the Rich render process before generating text with ANSI codes."
-        )
-        console.print("\nConsider the following code:\n")
-        console.print(Syntax(code, "python", line_numbers=True))
-        console.print()
-        console.print(
-            "When you call [b]print()[/b], Rich [i]renders[/i] the object in to the the following:\n"
-        )
-        fragments = list(console.render(text))
-        console.print(fragments)
-        console.print()
-        console.print(
-            "The Segments are then processed to produce the following output:\n"
-        )
-        console.print(text)
-        console.print(
-            "\nYou will only need to know this if you are implementing your own Rich renderables."
-        )
+    console.rule("rich.Segment")
+    console.print(
+        "A Segment is the last step in the Rich render process before generating text with ANSI codes."
+    )
+    console.print("\nConsider the following code:\n")
+    console.print(Syntax(code, "python", line_numbers=True))
+    console.print()
+    console.print(
+        "When you call [b]print()[/b], Rich [i]renders[/i] the object in to the the following:\n"
+    )
+    fragments = list(console.render(text))
+    console.print(fragments)
+    console.print()
+    console.print("The Segments are then processed to produce the following output:\n")
+    console.print(text)
+    console.print(
+        "\nYou will only need to know this if you are implementing your own Rich renderables."
+    )
